PHREEQC should now have no USE_MPI
Also no MERGE_FILES in PHREEQC

Not sure about deallocates in worker_closef

git-svn-id: svn://136.177.114.72/svn_GW/phreeqc3/trunk@7045 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/phreeqc/inverse.cpp
@@ -134,7 +134,7 @@ inverse_models(void)
 		}
 	}
 
-	user_punch_count_headings = old_headings.size();
+	user_punch_count_headings = (int) old_headings.size();
 	user_punch_headings = (const char **) PHRQ_realloc(user_punch_headings,
 		(size_t) (user_punch_count_headings + 1) * sizeof(char *));
 	if (user_punch_headings == NULL)
@@ -2125,7 +2125,7 @@ punch_model_heading(struct inverse *inv_ptr)
 
 	size_t j;
 
-	user_punch_count_headings = heading_names.size();
+	user_punch_count_headings = (int) heading_names.size();
 	user_punch_headings = (const char **) PHRQ_realloc(user_punch_headings,
 		(size_t) (user_punch_count_headings + 1) * sizeof(char *));
 	if (user_punch_headings == NULL)
